+++ packages/fluentui/docs/src/components/ComponentDoc/ComponentPerfExample/ComponentPerfChart.tsx
@@ -90,7 +90,7 @@ export const ComponentPerfChart = ({ perfTestName }) => {
         {process.env.NODE_ENV !== 'production' && (
           <RadioGroup
             defaultCheckedValue={FILTER_BY.CI_BUILD}
-            checkedValueChanged={handleFilterChange}
+            onCheckedValueChange={handleFilterChange}
             items={[
               { key: 'ci-build', label: FILTER_BY.CI_BUILD, value: FILTER_BY.CI_BUILD },
               { key: 'release', label: FILTER_BY.RELEASE, value: FILTER_BY.RELEASE },