+++ packages/fluentui/docs/src/examples/components/RadioGroup/Types/RadioGroupExample.shorthand.tsx
@@ -1,22 +1,18 @@
-import * as React from 'react'
-import { Divider, RadioGroup, Input, Text } from '@fluentui/react'
+import * as React from 'react';
+import { Divider, RadioGroup, Input, Text } from '@fluentui/react';
 
 class RadioGroupVerticalExample extends React.Component {
-  state = { selectedValue: '', inputTabIndex: '-1' }
+  state = { selectedValue: '', inputTabIndex: '-1' };
 
   render() {
-    const { selectedValue } = this.state
+    const { selectedValue } = this.state;
     return (
       <div>
         The selected value is: {selectedValue}
         <Divider />
-        <RadioGroup
-          defaultCheckedValue="capricciosa"
-          items={this.getItems()}
-          checkedValueChanged={this.handleChange}
-        />
+        <RadioGroup defaultCheckedValue="capricciosa" items={this.getItems()} onCheckedValueChange={this.handleChange} />
       </div>
-    )
+    );
   }
 
   getItems() {
@@ -27,28 +23,26 @@ class RadioGroupVerticalExample extends React.Component {
         key: 'Prosciutto',
         label: 'Prosciutto',
         value: 'prosciutto',
-        disabled: true,
+        disabled: true
       },
       {
         name: 'pizza',
         key: 'Custom',
         label: (
           <Text>
-            Choose your own{' '}
-            <Input input={{ tabIndex: this.state.inputTabIndex }} inline placeholder="flavour" />
+            Choose your own <Input input={{ tabIndex: this.state.inputTabIndex }} inline placeholder="flavour" />
           </Text>
         ),
         value: 'custom',
         checkedChanged: this.handleCustomCheckedChange,
-        'aria-label': 'Press Tab to change flavour',
-      },
-    ]
+        'aria-label': 'Press Tab to change flavour'
+      }
+    ];
   }
 
-  handleChange = (e, props) => this.setState({ selectedValue: props.value })
+  handleChange = (e, props) => this.setState({ selectedValue: props.value });
 
-  handleCustomCheckedChange = (e, props) =>
-    this.setState({ inputTabIndex: props.checked ? '0' : '-1' })
+  handleCustomCheckedChange = (e, props) => this.setState({ inputTabIndex: props.checked ? '0' : '-1' });
 }
 
-export default RadioGroupVerticalExample
+export default RadioGroupVerticalExample;