+++ packages/fluentui/docs/src/examples/components/RadioGroup/Types/RadioGroupVerticalExample.shorthand.tsx
@@ -1,23 +1,18 @@
-import * as React from 'react'
-import { Divider, RadioGroup, Input, Text } from '@fluentui/react'
+import * as React from 'react';
+import { Divider, RadioGroup, Input, Text } from '@fluentui/react';
 
 class RadioGroupVerticalExample extends React.Component {
-  state = { selectedValue: '', inputTabIndex: '-1' }
+  state = { selectedValue: '', inputTabIndex: '-1' };
 
   render() {
-    const { selectedValue } = this.state
+    const { selectedValue } = this.state;
     return (
       <div style={{ maxWidth: '400px' }}>
         The selected value is: {selectedValue}
         <Divider />
-        <RadioGroup
-          vertical
-          defaultCheckedValue="prosciutto"
-          items={this.getItems()}
-          checkedValueChanged={this.handleChange}
-        />
+        <RadioGroup vertical defaultCheckedValue="prosciutto" items={this.getItems()} onCheckedValueChange={this.handleChange} />
       </div>
-    )
+    );
   }
 
   getItems() {
@@ -26,20 +21,20 @@ class RadioGroupVerticalExample extends React.Component {
         name: 'pizza',
         key: 'Capricciosa',
         label: 'Capricciosa',
-        value: 'capricciosa',
+        value: 'capricciosa'
       },
       {
         name: 'pizza',
         key: 'Prosciutto',
         label: 'Prosciutto',
         value: 'prosciutto',
-        disabled: true,
+        disabled: true
       },
       {
         name: 'pizza',
         key: 'Pepperoni',
         label: 'Pepperoni',
-        value: 'pepperoni',
+        value: 'pepperoni'
       },
       {
         name: 'pizza',
@@ -53,8 +48,8 @@ class RadioGroupVerticalExample extends React.Component {
               styles={{
                 fontSize: '12px',
                 '& .ui-input__input': {
-                  padding: '2px 8px',
-                },
+                  padding: '2px 8px'
+                }
               }}
               placeholder="flavour"
             />
@@ -62,15 +57,14 @@ class RadioGroupVerticalExample extends React.Component {
         ),
         value: 'custom',
         checkedChanged: this.handleCustomCheckedChange,
-        'aria-label': 'Press Tab to change flavour',
-      },
-    ]
+        'aria-label': 'Press Tab to change flavour'
+      }
+    ];
   }
 
-  handleChange = (e, props) => this.setState({ selectedValue: props.value })
+  handleChange = (e, props) => this.setState({ selectedValue: props.value });
 
-  handleCustomCheckedChange = (e, props) =>
-    this.setState({ inputTabIndex: props.checked ? '0' : '-1' })
+  handleCustomCheckedChange = (e, props) => this.setState({ inputTabIndex: props.checked ? '0' : '-1' });
 }
 
-export default RadioGroupVerticalExample
+export default RadioGroupVerticalExample;