+++ packages/fluentui/docs/src/views/CategoryColorSchemes.tsx
@@ -1,15 +1,15 @@
-import * as React from 'react'
-import DocPage from '../components/DocPage/DocPage'
-import GuidesNavigationFooter from '../components/GuidesNavigationFooter'
-import CategoryColorSchemes from '../components/CategoryColorSchemes'
+import * as React from 'react';
+import DocPage from '../components/DocPage/DocPage';
+import GuidesNavigationFooter from '../components/GuidesNavigationFooter';
+import CategoryColorSchemes from '../components/CategoryColorSchemes';
 
-import { Dropdown, themes, Flex, Provider } from '@fluentui/react'
-import { faderStyles } from '../components/Fader'
-import { colorVariantsStyles } from '../components/ColorVariants'
-import { colorBoxStyles, colorBoxVariables } from '../components/ColorBox'
+import { Dropdown, themes, Flex, Provider } from '@fluentui/react';
+import { faderStyles } from '../components/Fader';
+import { colorVariantsStyles } from '../components/ColorVariants';
+import { colorBoxStyles, colorBoxVariables } from '../components/ColorBox';
 
 export default () => {
-  const [color, setColor] = React.useState('red')
+  const [color, setColor] = React.useState('red');
   return (
     <Provider
       theme={{
@@ -19,13 +19,13 @@ export default () => {
           Fader: faderStyles,
           Header: {
             root: {
-              fontWeight: 700,
-            },
-          },
+              fontWeight: 700
+            }
+          }
         },
         componentVariables: {
-          ColorBox: colorBoxVariables,
-        },
+          ColorBox: colorBoxVariables
+        }
       }}
     >
       <DocPage title="Category color schemes">
@@ -43,31 +43,31 @@ export default () => {
               'oliveDark',
               'olive',
               'neon',
-              'formatting',
+              'formatting'
             ]}
             defaultValue={'red'}
             placeholder="Select the color"
-            onSelectedChange={(e, { value }) => setColor(value as string)}
+            onChange={(e, { value }) => setColor(value as string)}
           />
           <CategoryColorSchemes
             themes={[themes.teams, themes.teamsHighContrast, themes.teamsDark]}
             headers={[
               {
                 as: 'h3',
-                content: 'Design token',
+                content: 'Design token'
               },
               {
                 as: 'h3',
-                content: 'Light theme',
+                content: 'Light theme'
               },
               {
                 as: 'h3',
-                content: 'HC theme',
+                content: 'HC theme'
               },
               {
                 as: 'h3',
-                content: 'Dark theme',
-              },
+                content: 'Dark theme'
+              }
             ]}
             name={color}
           />
@@ -75,5 +75,5 @@ export default () => {
         <GuidesNavigationFooter previous={{ name: 'Colors', url: 'colors' }} />
       </DocPage>
     </Provider>
-  )
-}
+  );
+};