+++ packages/fluentui/react-bindings/src/hooks/useStyles.ts
@@ -34,6 +34,7 @@ type InlineStyleProps<StyleProps> = {
 const defaultContext: StylesContextValue<{ renderRule: RendererRenderRule }> = {
   disableAnimations: false,
   performance: {
+    enableSanitizeCssPlugin: process.env.NODE_ENV !== 'production',
     enableStylesCaching: true,
     enableVariablesCaching: true
   },
@@ -43,8 +44,6 @@ const defaultContext: StylesContextValue<{ renderRule: RendererRenderRule }> = {
 
 const useStyles = <StyleProps extends PrimitiveProps>(displayName: string, options: UseStylesOptions<StyleProps>): UseStylesResult => {
   const context: StylesContextValue<{ renderRule: RendererRenderRule }> = React.useContext(ThemeContext) || defaultContext;
-
-  const { enableStylesCaching = true, enableVariablesCaching = true } = context.performance || {};
 
   const {
     className = process.env.NODE_ENV === 'production' ? '' : 'no-classname-🙉',
@@ -70,10 +69,7 @@ const useStyles = <StyleProps extends PrimitiveProps>(displayName: string, optio
     rtl,
     saveDebug: fluentUIDebug => (debug.current = { fluentUIDebug }),
     theme: context.theme,
-    performance: {
-      enableStylesCaching,
-      enableVariablesCaching
-    }
+    performance: context.performance
   });
 
   return { classes, styles: resolvedStyles };