+++ packages/fluentui/react/src/components/Dropdown/Dropdown.tsx
@@ -146,6 +146,20 @@ export interface DropdownProps extends UIComponentProps<DropdownProps, DropdownS
   noResultsMessage?: ShorthandValue<ListItemProps>;
 
   /**
+   * Called when the dropdown's selected items index change.
+   * @param event - React's original SyntheticEvent.
+   * @param data - All props and the new selected value(s).
+   */
+  onActiveSelectedIndexChange?: ComponentEventHandler<DropdownProps>;
+
+  /**
+   * Called when the dropdown's highlighted index change.
+   * @param event - React's original SyntheticEvent.
+   * @param data - All props and the new selected value(s).
+   */
+  onHighlightedIndexChange?: ComponentEventHandler<DropdownProps>;
+
+  /**
    * Called when the dropdown opens or closes.
    * @param event - React's original SyntheticEvent.
    * @param data - All props, with `open` reflecting the new open state.
@@ -164,7 +178,7 @@ export interface DropdownProps extends UIComponentProps<DropdownProps, DropdownS
    * @param event - React's original SyntheticEvent.
    * @param data - All props and the new selected value(s).
    */
-  onSelectedChange?: ComponentEventHandler<DropdownProps>;
+  onChange?: ComponentEventHandler<DropdownProps>;
 
   /** A dropdown's open state can be controlled. */
   open?: boolean;
@@ -275,7 +289,9 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
     offset: PropTypes.string,
     onOpenChange: PropTypes.func,
     onSearchQueryChange: PropTypes.func,
-    onSelectedChange: PropTypes.func,
+    onChange: PropTypes.func,
+    onActiveSelectedIndexChange: PropTypes.func,
+    onHighlightedIndexChange: PropTypes.func,
     open: PropTypes.bool,
     placeholder: PropTypes.string,
     position: PropTypes.oneOf(POSITIONS),
@@ -384,12 +400,18 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
     return modifiedState;
   }
 
+  handleChange = (e: React.SyntheticEvent) => {
+    // Dropdown component doesn't present any `input` component in markup, however all of our
+    // components should handle events transparently.
+    _.invoke(this.props, 'onChange', e, { ...this.props, value: this.state.value });
+  };
+
   renderComponent({ ElementType, classes, styles, variables, unhandledProps, rtl }: RenderResultConfig<DropdownProps>) {
     const { clearable, clearIndicator, search, multiple, getA11yStatusMessage, itemToString, toggleIndicator } = this.props;
     const { highlightedIndex, open, searchQuery, value } = this.state;
 
     return (
-      <ElementType className={classes.root} {...unhandledProps}>
+      <ElementType className={classes.root} onChange={this.handleChange} {...unhandledProps}>
         <Downshift
           isOpen={open}
           onChange={this.handleSelectedChange}
@@ -732,7 +754,7 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
   }
 
   handleSearchQueryChange = (searchQuery: string) => {
-    this.setStateAndInvokeHandler('onSearchQueryChange', null, {
+    this.setStateAndInvokeHandler(['onSearchQueryChange', 'onHighlightedIndexChange'], null, {
       searchQuery,
       highlightedIndex: this.props.highlightFirstItemOnOpen ? 0 : null,
       open: searchQuery === '' ? false : this.state.open
@@ -773,12 +795,15 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
         newState.highlightedIndex = null;
       }
 
-      this.setStateAndInvokeHandler('onOpenChange', null, newState);
+      this.setStateAndInvokeHandler(['onOpenChange'], null, newState);
     }
 
     if (this.state.open && _.isNumber(changes.highlightedIndex)) {
       const itemIsFromKeyboard = changes.type !== Downshift.stateChangeTypes.itemMouseEnter;
-      this.setState({ itemIsFromKeyboard, highlightedIndex: changes.highlightedIndex });
+      this.setStateAndInvokeHandler(['onHighlightedIndexChange'], null, {
+        itemIsFromKeyboard,
+        highlightedIndex: changes.highlightedIndex
+      });
     }
   };
 
@@ -816,8 +841,9 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
       this.handleSelectedItemRemove(e, item, predefinedProps, dropdownSelectedItemProps);
     },
     onClick: (e: React.SyntheticEvent, dropdownSelectedItemProps: DropdownSelectedItemProps) => {
-      this.setState({ activeSelectedIndex: this.state.value.indexOf(item) });
-
+      this.setStateAndInvokeHandler(['onActiveSelectedIndexChange'], null, {
+        activeSelectedIndex: this.state.value.indexOf(item)
+      });
       e.stopPropagation();
       _.invoke(predefinedProps, 'onClick', e, dropdownSelectedItemProps);
     },
@@ -932,10 +958,14 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
       // If last element was already active, perform a 'reset' of activeSelectedIndex.
       if (this.state.activeSelectedIndex === value.length - 1) {
         this.setState({ activeSelectedIndex: null }, () => {
-          this.setState({ activeSelectedIndex: value.length - 1 });
+          this.setStateAndInvokeHandler(['onActiveSelectedIndexChange'], null, {
+            activeSelectedIndex: value.length - 1
+          });
         });
       } else {
-        this.setState({ activeSelectedIndex: value.length - 1 });
+        this.setStateAndInvokeHandler(['onActiveSelectedIndexChange'], null, {
+          activeSelectedIndex: value.length - 1
+        });
       }
     }
   };
@@ -953,14 +983,13 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
   handleClear = (e: React.SyntheticEvent<HTMLElement>) => {
     const { activeSelectedIndex, highlightedIndex, open, searchQuery, value } = this.getInitialAutoControlledState(this.props);
 
-    this.setStateAndInvokeHandler('onSelectedChange', e, {
+    this.setStateAndInvokeHandler(['onChange', 'onActiveSelectedIndexChange', 'onHighlightedIndexChange'], e, {
       activeSelectedIndex,
       highlightedIndex,
       open,
       searchQuery,
       value
     });
-    this.setState({ activeSelectedIndex, highlightedIndex, open, searchQuery, value });
 
     this.tryFocusSearchInput();
     this.tryFocusTriggerButton();
@@ -1018,13 +1047,15 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
     const { items, multiple, getA11ySelectionMessage } = this.props;
     const { value } = this.state;
 
-    this.setStateAndInvokeHandler('onSelectedChange', null, {
+    this.setStateAndInvokeHandler(['onChange'], null, {
       searchQuery: this.getSelectedItemAsString(item),
       value: multiple ? [...value, item] : [item]
     });
 
-    if (!multiple) {
-      this.setState({ highlightedIndex: items.indexOf(item) });
+    if (!multiple && items) {
+      this.setStateAndInvokeHandler(['onHighlightedIndexChange'], null, {
+        highlightedIndex: items.indexOf(item)
+      });
     }
 
     if (getA11ySelectionMessage && getA11ySelectionMessage.onAdd) {
@@ -1057,7 +1088,9 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
         break;
       case previousKey:
         if (value.length > 0 && !_.isNil(activeSelectedIndex) && activeSelectedIndex > 0) {
-          this.setState({ activeSelectedIndex: activeSelectedIndex - 1 });
+          this.setStateAndInvokeHandler(['onActiveSelectedIndexChange'], null, {
+            activeSelectedIndex: activeSelectedIndex - 1
+          });
         }
         break;
       case nextKey:
@@ -1123,7 +1156,7 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
     }
 
     if (newHighlightedIndex >= 0) {
-      this.setState({
+      this.setStateAndInvokeHandler(['onHighlightedIndexChange'], null, {
         highlightedIndex: newHighlightedIndex
       });
     }
@@ -1135,7 +1168,9 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
     predefinedProps: DropdownSelectedItemProps,
     dropdownSelectedItemProps: DropdownSelectedItemProps
   ) {
-    this.setState({ activeSelectedIndex: null });
+    this.setStateAndInvokeHandler(['onActiveSelectedIndexChange'], null, {
+      activeSelectedIndex: null
+    });
     this.removeItemFromValue(item);
     this.tryFocusSearchInput();
     this.tryFocusTriggerButton();
@@ -1157,7 +1192,7 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
       this.setA11ySelectionMessage(getA11ySelectionMessage.onRemove(poppedItem));
     }
 
-    this.setStateAndInvokeHandler('onSelectedChange', null, { value });
+    this.setStateAndInvokeHandler(['onChange'], null, { value });
   }
 
   /**
@@ -1166,7 +1201,7 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
    * because we want to keep the event handling interface
    */
   setStateAndInvokeHandler = (
-    handlerName: keyof DropdownProps,
+    handlerNames: (keyof DropdownProps)[],
     event: React.SyntheticEvent<HTMLElement>,
     newState: Partial<DropdownState>
   ) => {
@@ -1176,7 +1211,9 @@ class Dropdown extends AutoControlledComponent<WithAsProp<DropdownProps>, Dropdo
     const newValue = this.props.multiple ? proposedValue : proposedValue[0] || null;
 
     this.setState(newState as DropdownState);
-    _.invoke(this.props, handlerName, event, { ...this.props, ...newState, value: newValue });
+    handlerNames.forEach(handlerName => {
+      _.invoke(this.props, handlerName, event, { ...this.props, ...newState, value: newValue });
+    });
   };
 
   tryFocusTriggerButton = () => {