+++ packages/fluentui/react/src/components/Tree/Tree.tsx
@@ -1,13 +1,13 @@
-import { Accessibility, treeBehavior } from '@fluentui/accessibility'
-import { ReactAccessibilityBehavior, getNextElement } from '@fluentui/react-bindings'
-import * as customPropTypes from '@fluentui/react-proptypes'
-import * as _ from 'lodash'
-import * as PropTypes from 'prop-types'
-import * as React from 'react'
-import { Ref } from '@fluentui/react-component-ref'
+import { Accessibility, treeBehavior } from '@fluentui/accessibility';
+import { ReactAccessibilityBehavior, getNextElement } from '@fluentui/react-bindings';
+import * as customPropTypes from '@fluentui/react-proptypes';
+import * as _ from 'lodash';
+import * as PropTypes from 'prop-types';
+import * as React from 'react';
+import { Ref } from '@fluentui/react-component-ref';
 
-import TreeItem, { TreeItemProps } from './TreeItem'
-import TreeTitle, { TreeTitleProps } from './TreeTitle'
+import TreeItem, { TreeItemProps } from './TreeItem';
+import TreeTitle, { TreeTitleProps } from './TreeTitle';
 import {
   childrenExist,
   commonPropTypes,
@@ -17,36 +17,37 @@ import {
   rtlTextContainer,
   applyAccessibilityKeyHandlers,
   AutoControlledComponent,
-  ShorthandFactory,
-} from '../../utils'
+  ShorthandFactory
+} from '../../utils';
 import {
   ShorthandRenderFunction,
   WithAsProp,
   withSafeTypeForAs,
   ShorthandCollection,
   ShorthandValue,
-} from '../../types'
-import { hasSubtree, removeItemAtIndex } from './utils'
+  ComponentEventHandler
+} from '../../types';
+import { hasSubtree, removeItemAtIndex } from './utils';
 
 export interface TreeSlotClassNames {
-  item: string
+  item: string;
 }
 
 export interface TreeProps extends UIComponentProps, ChildrenComponentProps {
   /** Accessibility behavior if overridden by the user. */
-  accessibility?: Accessibility
+  accessibility?: Accessibility;
 
   /** Ids of expanded items. */
-  activeItemIds?: string[]
+  activeItemIds?: string[];
 
   /** Initial activeItemIds value. */
-  defaultActiveItemIds?: string[]
+  defaultActiveItemIds?: string[];
 
   /** Only allow one subtree to be expanded at a time. */
-  exclusive?: boolean
+  exclusive?: boolean;
 
   /** Shorthand array of props for Tree. */
-  items?: ShorthandCollection<TreeItemProps>
+  items?: ShorthandCollection<TreeItemProps>;
 
   /**
    * A custom render function for the title slot.
@@ -55,7 +56,14 @@ export interface TreeProps extends UIComponentProps, ChildrenComponentProps {
    * @param props - The computed props for this slot.
    * @param children - The computed children for this slot.
    */
-  renderItemTitle?: ShorthandRenderFunction<TreeTitleProps>
+  renderItemTitle?: ShorthandRenderFunction<TreeTitleProps>;
+
+  /**
+   * Called when active item ids change.
+   * @param event - React's original SyntheticEvent.
+   * @param data - All props, with `activeItemIds` reflecting the new state.
+   */
+  onActiveItemIdsChange?: ComponentEventHandler<TreeProps>;
 
   /**
    * Callback that provides rendered tree items to be used by react-virtualized for instance.
@@ -64,274 +72,262 @@ export interface TreeProps extends UIComponentProps, ChildrenComponentProps {
    * @param renderedItem - The array of rendered items.
    * @returns The render prop result.
    */
-  renderedItems?: (renderedItems: React.ReactElement[]) => React.ReactNode
+  renderedItems?: (renderedItems: React.ReactElement[]) => React.ReactNode;
 }
 
 export interface TreeItemForRenderProps {
-  elementRef: React.RefObject<HTMLElement>
-  id: string
-  index: number
-  level: number
-  parent: ShorthandValue<TreeItemProps>
-  siblings: ShorthandCollection<TreeItemProps>
+  elementRef: React.RefObject<HTMLElement>;
+  id: string;
+  index: number;
+  level: number;
+  parent: ShorthandValue<TreeItemProps>;
+  siblings: ShorthandCollection<TreeItemProps>;
 }
 
 export interface TreeState {
-  activeItemIds: string[]
-  itemsForRender: Record<string, TreeItemForRenderProps>
+  activeItemIds: string[];
+  itemsForRender: Record<string, TreeItemForRenderProps>;
 }
 
 class Tree extends AutoControlledComponent<WithAsProp<TreeProps>, TreeState> {
-  static create: ShorthandFactory<TreeProps>
+  static create: ShorthandFactory<TreeProps>;
 
-  static displayName = 'Tree'
+  static displayName = 'Tree';
 
-  static className = 'ui-tree'
+  static className = 'ui-tree';
 
   static slotClassNames: TreeSlotClassNames = {
-    item: `${Tree.className}__item`,
-  }
+    item: `${Tree.className}__item`
+  };
 
   static propTypes = {
     ...commonPropTypes.createCommon({
-      content: false,
+      content: false
     }),
     activeItemIds: customPropTypes.collectionShorthand,
     defaultActiveItemIds: customPropTypes.collectionShorthand,
     exclusive: PropTypes.bool,
     items: customPropTypes.collectionShorthand,
+    onActiveItemIdsChange: PropTypes.func,
     renderItemTitle: PropTypes.func,
-    renderedItems: PropTypes.func,
-  }
+    renderedItems: PropTypes.func
+  };
 
   static defaultProps = {
     as: 'div',
-    accessibility: treeBehavior as Accessibility,
-  }
+    accessibility: treeBehavior as Accessibility
+  };
 
-  static autoControlledProps = ['activeItemIds']
+  static autoControlledProps = ['activeItemIds'];
 
-  static Item = TreeItem
-  static Title = TreeTitle
+  static Item = TreeItem;
+  static Title = TreeTitle;
 
   // memoize this function if performance issue occurs.
   static getItemsForRender = (itemsFromProps: ShorthandCollection<TreeItemProps>) => {
-    const itemsForRenderGenerator = (
-      items = itemsFromProps,
-      level = 1,
-      parent?: ShorthandValue<TreeItemProps>,
-    ) => {
+    const itemsForRenderGenerator = (items = itemsFromProps, level = 1, parent?: ShorthandValue<TreeItemProps>) => {
       return _.reduce(
         items,
         (acc: Object, item: ShorthandValue<TreeItemProps>, index: number) => {
-          const id = item['id']
-          const isSubtree = hasSubtree(item)
+          const id = item['id'];
+          const isSubtree = hasSubtree(item);
 
           acc[id] = {
             elementRef: React.createRef<HTMLElement>(),
             level,
             index: index + 1, // Used for aria-posinset and it's 1-based.
             parent,
-            siblings: items.filter(currentItem => currentItem !== item),
-          }
+            siblings: items.filter(currentItem => currentItem !== item)
+          };
 
           return {
             ...acc,
-            ...(isSubtree ? itemsForRenderGenerator(item['items'], level + 1, item) : {}),
-          }
+            ...(isSubtree ? itemsForRenderGenerator(item['items'], level + 1, item) : {})
+          };
         },
-        {},
-      )
-    }
+        {}
+      );
+    };
 
-    return itemsForRenderGenerator(itemsFromProps)
-  }
+    return itemsForRenderGenerator(itemsFromProps);
+  };
 
   static getAutoControlledStateFromProps(nextProps: TreeProps, prevState: TreeState) {
-    const { items } = nextProps
-    const itemsForRender = Tree.getItemsForRender(items)
-    let { activeItemIds } = nextProps
+    const { items } = nextProps;
+    const itemsForRender = Tree.getItemsForRender(items);
+    let { activeItemIds } = nextProps;
 
     if (!activeItemIds && items) {
-      activeItemIds = prevState.activeItemIds
+      activeItemIds = prevState.activeItemIds;
 
       const expandedItemsGenerator = (items, acc = activeItemIds) =>
         _.reduce(
           items,
           (acc, item) => {
             if (item['expanded'] && acc.indexOf(item['id']) === -1) {
-              acc.push(item['id'])
+              acc.push(item['id']);
             }
 
             if (item['items']) {
-              return expandedItemsGenerator(item['items'], acc)
+              return expandedItemsGenerator(item['items'], acc);
             }
 
-            return acc
+            return acc;
           },
-          acc,
-        )
+          acc
+        );
 
-      expandedItemsGenerator(items)
+      expandedItemsGenerator(items);
     }
 
     return {
       activeItemIds,
-      itemsForRender,
-    }
+      itemsForRender
+    };
   }
 
   getInitialAutoControlledState() {
-    return { activeItemIds: [] }
+    return { activeItemIds: [] };
   }
 
-  treeRef = React.createRef<HTMLElement>()
+  treeRef = React.createRef<HTMLElement>();
 
   handleTreeItemOverrides = (predefinedProps: TreeItemProps) => ({
     onTitleClick: (e: React.SyntheticEvent, treeItemProps: TreeItemProps) => {
       if (!hasSubtree(treeItemProps)) {
-        return
+        return;
       }
 
-      let { activeItemIds } = this.state
-      const { id, siblings } = treeItemProps
-      const { exclusive } = this.props
+      let { activeItemIds } = this.state;
+      const { id, siblings } = treeItemProps;
+      const { exclusive } = this.props;
 
-      const activeItemIdIndex = activeItemIds.indexOf(id)
+      const activeItemIdIndex = activeItemIds.indexOf(id);
 
       if (activeItemIdIndex > -1) {
-        activeItemIds = removeItemAtIndex(activeItemIds, activeItemIdIndex)
+        activeItemIds = removeItemAtIndex(activeItemIds, activeItemIdIndex);
       } else {
         if (exclusive) {
           siblings.some(sibling => {
-            const activeSiblingIdIndex = activeItemIds.indexOf(sibling['id'])
+            const activeSiblingIdIndex = activeItemIds.indexOf(sibling['id']);
             if (activeSiblingIdIndex > -1) {
-              activeItemIds = removeItemAtIndex(activeItemIds, activeSiblingIdIndex)
+              activeItemIds = removeItemAtIndex(activeItemIds, activeSiblingIdIndex);
 
-              return true
+              return true;
             }
-            return false
-          })
+            return false;
+          });
         }
 
-        activeItemIds = [...activeItemIds, id]
+        activeItemIds = [...activeItemIds, id];
       }
 
-      this.setState({
-        activeItemIds,
-      })
+      this.setActiveItemIds(e, activeItemIds);
 
-      _.invoke(predefinedProps, 'onTitleClick', e, treeItemProps)
+      _.invoke(predefinedProps, 'onTitleClick', e, treeItemProps);
     },
     onFocusParent: (e: React.SyntheticEvent, treeItemProps: TreeItemProps) => {
-      const { parent } = treeItemProps
+      const { parent } = treeItemProps;
 
       if (!parent) {
-        return
+        return;
       }
 
-      const { itemsForRender } = this.state
-      const parentItemForRender = itemsForRender[parent['id']]
+      const { itemsForRender } = this.state;
+      const parentItemForRender = itemsForRender[parent['id']];
 
-      if (
-        !parentItemForRender ||
-        !parentItemForRender.elementRef ||
-        !parentItemForRender.elementRef.current
-      ) {
-        return
+      if (!parentItemForRender || !parentItemForRender.elementRef || !parentItemForRender.elementRef.current) {
+        return;
       }
 
-      parentItemForRender.elementRef.current.focus()
-      _.invoke(predefinedProps, 'onFocusParent', e, treeItemProps)
+      parentItemForRender.elementRef.current.focus();
+      _.invoke(predefinedProps, 'onFocusParent', e, treeItemProps);
     },
     onFocusFirstChild: (e: React.SyntheticEvent, treeItemProps: TreeItemProps) => {
-      const { id } = treeItemProps
+      const { id } = treeItemProps;
 
-      const { itemsForRender } = this.state
-      const currentElement = itemsForRender[id].elementRef
+      const { itemsForRender } = this.state;
+      const currentElement = itemsForRender[id].elementRef;
 
       if (!currentElement || !currentElement.current) {
-        return
+        return;
       }
 
-      const elementToBeFocused = getNextElement(this.treeRef.current, currentElement.current)
+      const elementToBeFocused = getNextElement(this.treeRef.current, currentElement.current);
 
       if (!elementToBeFocused) {
-        return
+        return;
       }
 
-      elementToBeFocused.focus()
-      _.invoke(predefinedProps, 'onFocusFirstChild', e, treeItemProps)
+      elementToBeFocused.focus();
+      _.invoke(predefinedProps, 'onFocusFirstChild', e, treeItemProps);
     },
     onSiblingsExpand: (e: React.SyntheticEvent, treeItemProps: TreeItemProps) => {
       if (this.props.exclusive) {
-        return
+        return;
       }
 
-      const { id, siblings } = treeItemProps
-      const { activeItemIds } = this.state
+      const { id, siblings } = treeItemProps;
+      const { activeItemIds } = this.state;
 
       siblings.forEach(sibling => {
         if (hasSubtree(sibling) && !this.isActiveItem(sibling['id'])) {
-          activeItemIds.push(sibling['id'])
+          activeItemIds.push(sibling['id']);
         }
-      })
+      });
 
       if (hasSubtree(treeItemProps) && !this.isActiveItem(id)) {
-        activeItemIds.push(id)
+        activeItemIds.push(id);
       }
 
-      this.setState({
-        activeItemIds,
-      })
+      this.setActiveItemIds(e, activeItemIds);
 
-      _.invoke(predefinedProps, 'onSiblingsExpand', e, treeItemProps)
-    },
-  })
+      _.invoke(predefinedProps, 'onSiblingsExpand', e, treeItemProps);
+    }
+  });
+
+  setActiveItemIds = (e: React.SyntheticEvent, activeItemIds: string[]) => {
+    _.invoke(this.props, 'onActiveItemIds', e, { ...this.props, activeItemIds });
+
+    this.setState({
+      activeItemIds
+    });
+  };
 
   renderContent(accessibility: ReactAccessibilityBehavior): React.ReactElement[] {
-    const { itemsForRender } = this.state
-    const { items, renderItemTitle } = this.props
+    const { itemsForRender } = this.state;
+    const { items, renderItemTitle } = this.props;
 
-    if (!items) return null
+    if (!items) return null;
 
     const renderItems = (items: ShorthandCollection<TreeItemProps>): React.ReactElement[] => {
-      return items.reduce(
-        (renderedItems: React.ReactElement[], item: ShorthandValue<TreeItemProps>) => {
-          const itemForRender = itemsForRender[item['id']]
-          const { elementRef, ...restItemForRender } = itemForRender
-          const isSubtree = hasSubtree(item)
-          const isSubtreeExpanded = isSubtree && this.isActiveItem(item['id'])
-          const renderedItem = TreeItem.create(item, {
-            defaultProps: () => ({
-              accessibility: accessibility.childBehaviors
-                ? accessibility.childBehaviors.item
-                : undefined,
-              className: Tree.slotClassNames.item,
-              expanded: isSubtreeExpanded,
-              renderItemTitle,
-              key: item['id'],
-              contentRef: elementRef,
-              ...restItemForRender,
-            }),
-            overrideProps: this.handleTreeItemOverrides,
-          })
+      return items.reduce((renderedItems: React.ReactElement[], item: ShorthandValue<TreeItemProps>) => {
+        const itemForRender = itemsForRender[item['id']];
+        const { elementRef, ...restItemForRender } = itemForRender;
+        const isSubtree = hasSubtree(item);
+        const isSubtreeExpanded = isSubtree && this.isActiveItem(item['id']);
+        const renderedItem = TreeItem.create(item, {
+          defaultProps: () => ({
+            accessibility: accessibility.childBehaviors ? accessibility.childBehaviors.item : undefined,
+            className: Tree.slotClassNames.item,
+            expanded: isSubtreeExpanded,
+            renderItemTitle,
+            key: item['id'],
+            contentRef: elementRef,
+            ...restItemForRender
+          }),
+          overrideProps: this.handleTreeItemOverrides
+        });
 
-          return [
-            ...renderedItems,
-            renderedItem,
-            ...(isSubtreeExpanded ? renderItems(item['items']) : ([] as any)),
-          ]
-        },
-        [],
-      )
-    }
+        return [...renderedItems, renderedItem, ...(isSubtreeExpanded ? renderItems(item['items']) : ([] as any))];
+      }, []);
+    };
 
-    return renderItems(items)
+    return renderItems(items);
   }
 
   renderComponent({ ElementType, classes, accessibility, unhandledProps }) {
-    const { children, renderedItems } = this.props
+    const { children, renderedItems } = this.props;
 
     return (
       <Ref innerRef={this.treeRef}>
@@ -349,19 +345,19 @@ class Tree extends AutoControlledComponent<WithAsProp<TreeProps>, TreeState> {
             : this.renderContent(accessibility)}
         </ElementType>
       </Ref>
-    )
+    );
   }
 
   isActiveItem = (id: string): boolean => {
-    const { activeItemIds } = this.state
-    return activeItemIds.indexOf(id) > -1
-  }
+    const { activeItemIds } = this.state;
+    return activeItemIds.indexOf(id) > -1;
+  };
 }
 
 Tree.create = createShorthandFactory({
   Component: Tree,
-  mappedArrayProp: 'items',
-})
+  mappedArrayProp: 'items'
+});
 
 /**
  * A Tree displays data organised in tree hierarchy.
@@ -370,6 +366,7 @@ Tree.create = createShorthandFactory({
  * Implements [ARIA TreeView](https://www.w3.org/TR/wai-aria-practices-1.1/#TreeView) design pattern.
  * @accessibilityIssues
  * [Treeview - JAWS doesn't narrate position for each tree item](https://github.com/FreedomScientific/VFO-standards-support/issues/338)
+ * [Aria compliant trees are read as empty tables](https://bugs.chromium.org/p/chromium/issues/detail?id=1048770)
  */
 
-export default withSafeTypeForAs<typeof Tree, TreeProps, 'ul'>(Tree)
+export default withSafeTypeForAs<typeof Tree, TreeProps, 'ul'>(Tree);