+++ packages/fluentui/react/test/specs/components/Embed/Embed-test.tsx
@@ -1,45 +1,39 @@
-import * as React from 'react'
-import Embed from 'src/components/Embed/Embed'
-import { isConformant, handlesAccessibility } from 'test/specs/commonTests'
-import { mountWithProviderAndGetComponent } from 'test/utils'
+import * as React from 'react';
+import Embed from 'src/components/Embed/Embed';
+import { isConformant, handlesAccessibility } from 'test/specs/commonTests';
+import { mountWithProviderAndGetComponent } from 'test/utils';
 
 describe('Embed', () => {
-  isConformant(Embed)
+  isConformant(Embed, { autoControlledProps: ['active'] });
 
   describe('accessibility', () => {
-    handlesAccessibility(Embed, { defaultRootRole: 'presentation' })
-  })
+    handlesAccessibility(Embed, { defaultRootRole: 'presentation' });
+  });
 
   describe('onClick', () => {
     test('is called with (e, props) on a click', () => {
-      const onClick = jest.fn()
-      const embed = mountWithProviderAndGetComponent(Embed, <Embed onClick={onClick} />)
+      const onClick = jest.fn();
+      const embed = mountWithProviderAndGetComponent(Embed, <Embed onClick={onClick} />);
 
-      embed.simulate('click')
+      embed.simulate('click');
 
-      expect(onClick).toHaveBeenCalledTimes(1)
-      expect(onClick).toHaveBeenCalledWith(
-        expect.objectContaining({ type: 'click' }),
-        expect.objectContaining({ onClick, active: true }),
-      )
-    })
-  })
+      expect(onClick).toHaveBeenCalledTimes(1);
+      expect(onClick).toHaveBeenCalledWith(expect.objectContaining({ type: 'click' }), expect.objectContaining({ onClick, active: true }));
+    });
+  });
 
-  describe('onActiveChanged', () => {
+  describe('onActiveChange', () => {
     test('is called with (e, props) on a click', () => {
-      const onActiveChanged = jest.fn()
-      const embed = mountWithProviderAndGetComponent(
-        Embed,
-        <Embed onActiveChanged={onActiveChanged} />,
-      )
+      const onActiveChange = jest.fn();
+      const embed = mountWithProviderAndGetComponent(Embed, <Embed onActiveChange={onActiveChange} />);
 
-      embed.simulate('click')
+      embed.simulate('click');
 
-      expect(onActiveChanged).toHaveBeenCalledTimes(1)
-      expect(onActiveChanged).toHaveBeenCalledWith(
+      expect(onActiveChange).toHaveBeenCalledTimes(1);
+      expect(onActiveChange).toHaveBeenCalledWith(
         expect.objectContaining({ type: 'click' }),
-        expect.objectContaining({ onActiveChanged, active: true }),
-      )
-    })
-  })
-})
+        expect.objectContaining({ onActiveChange, active: true })
+      );
+    });
+  });
+});