+++ packages/fluentui/react/test/specs/components/RadioGroup/RadioGroupItem-test.ts
@@ -1,13 +1,13 @@
-import { isConformant, handlesAccessibility } from 'test/specs/commonTests'
+import { isConformant, handlesAccessibility } from 'test/specs/commonTests';
 
-import RadioGroupItem from 'src/components/RadioGroup/RadioGroupItem'
+import RadioGroupItem from 'src/components/RadioGroup/RadioGroupItem';
 
 describe('RadioGroupItem', () => {
-  isConformant(RadioGroupItem)
+  isConformant(RadioGroupItem, { autoControlledProps: ['checked'] });
 
   describe('accessibility', () => {
     handlesAccessibility(RadioGroupItem, {
-      defaultRootRole: 'radio',
-    })
-  })
-})
+      defaultRootRole: 'radio'
+    });
+  });
+});